appveyor.yml unnecessarily complicated
In the appveyor.yml most string scalars (but not all) were double quoted scalars. YAML allows any string through backslashes within double quotes, but also requires backslashes and double quotes to be escaped. The example can be made much more readable
by using non-quoted scalar strings, or single quoted ones (in case the scalar starts with a `%` or can be interpreted as an integer.

This was originally proposed as change for the user guide (https://github.com/pypa/python-packaging-user-guide/issues/221#issuecomment-249680859)

(Depending on how the program that loads the file handles the value for `PYTHON_ARCH`, the quotes around those values could be removed  as well, assuming the program transparently converts from integer to string internally if necessary. )